♻️(front) move getFilterFromState to a utils folder
getFilterFromState is going to be reused in another component. Therefore
it is time to move it from being colocated with the only component
that used it to a shared utils folder.
