+++ richie/js/utils/filters/getFilterFromState.ts
